fix thread_rng arguments
The arkworks `rand` methods are defined as part of the `UniformRand`
trait, but the actual signature of the method just uses a `rand::Rng`
bound. That means we need to `use` the arkworks trait, but actually
bounding our generic functions by it serves no purpose but to confuse
the compiler.

Thanks to Anton Lazarev for help debugging this!